Move tests to correct folders
