+++ src/ggrc-client/js/models/custom-attributes/test/custom-attribute-definition_spec.js
