+++ src/ggrc-client/js/models/mixins/tests/access-control-list_spec.js
@@ -2,7 +2,8 @@
   Copyright (C) 2018 Google Inc.
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
-import * as snapshotUtils from '../../plugins/utils/snapshot-utils';
+
+import * as snapshotUtils from '../../../plugins/utils/snapshot-utils';
 
 describe('can.Model.Mixin.accessControlList', () => {
   describe('"cleanupAcl" method: ', () => {
