+++ src/ggrc-client/js/models/mixins/tests/related-assessments-loader_spec.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import {makeFakeInstance} from '../../../js_specs/spec_helpers';
+import {makeFakeInstance} from '../../../../js_specs/spec_helpers';
 
 const ENDPOINT = '/api/related_assessments';
